fix: consolidate duplicate AugmentationsDisplay test files (#514)
Merge 10 cyberlimb CRUD tests from top-level mirror file into the
co-located test file, eliminating mock drift that caused CI failures
when the component gained new imports. Add makeCyberlimb() factory
to shared test-helpers.
